Added event listeners for SeekBar and RadioButtons
The SeekBar can now update the tip percentage, and gets updated when the user presses -/+. The -/+ also disable and enable the decrease button if the user tries to enter a negative tip percentage.

The RadioGroup for number of people hides the Total per Person row if one person is selected.